test: consolidate config round-trip tests
